requirements.txt
# Kafka
kafka-python==2.0.2

# Apache Spark
pyspark==3.5.1

# Deep Learning & NLP
transformers==4.36.0
torch==2.1.2
sentencepiece==0.1.99

# Database
psycopg2-binary==2.9.9

# Configuration
pyyaml==6.0.1

# AWS S3
boto3==1.34.0

# Web Scraping
apify-client==1.7.0

# Utilities
python-dateutil==2.8.2

pandas==1.5.3
pyarrow>=4.0.0